[optimization] Cache pluralization
As it turns out, pluralizing "job" and "social_security" was
taking up 4% of the total simulator runtime. We can cut almost
all of that out just by caching that result. I also refactored
the method to use `flat_map`, which is slightly faster than a
map and flatten (or map and concat).