[optimization] Add acceptance test for accuracy
I added FortuneTeller::Benchmark as a reference point, and
extracted the code from the benchmark and profiling scripts
to it. Then, using Timecop, we can freeze the testing date
and assert that we get the same results each time.

This test will need to change when randomization enters the
picture, but for now it'll help ensure refactors don't break
the sim.