[holdings] Deal with holdings separately
This doesn't yet separate rates of returns, but it does accrue interest on,
debit, and credit holdings separately.

For now, we are using bare-bones allocation and distribution strategies.  For
allocations, the savings plan specifies which holding to entirely allocate new
savings to; for distributions, the plan just drains the smallest accounts
first. Both of these strategies will be revamped in the future, but for this
change I didn't spend much time on them.