[holdings] Implement wage growth
The value for wage growth is determined from growth rates, though note
that if the wage_growth value for year 2017 is 1.02, then the wages
for 2018 will be 1.02 times the wages of 2017 (assuming that wage growth
refers to how much wages grow during the given year). Otherwise, the
first value passed in would never be used.