[holdings] Fix cache inval for static transforms
Since wage growth and inflation are both factored into the transforms generated
by Job and SocialSecurity, the caching of those transforms must also depend on
the wage growth and inflation rates. To avoid invalidating the cache when
more-likely-to-change rates are changed (it's most likely we'll just play
around with rates of return, and leave wage growth largely fixed), the cache
key depends just on those two rate streams.

However, simulating a large number of scenarios with different wage growths
will cause a memory leak. We may wish to implement an LRU cache to avoid this,
or just avoid caching at all if that ends up being our algorithm.