[holdings] Remove #inflating_int from public API
Values being passed to set up the context of the simulation should be passed
primarily into DayPlan objects, so Simulator#inflating_int was moved down into
DayPlan#inflating_int. The only remaining caller, SocialSecurity::Generator,
uses a different start date, so I just replaced the call with a raw
initialization.

The ultimate benefit here is that the simulator is safe to inflate all values
on its own, as it sees fit, which leads to saner code down the line.