[holdings] Cache transforms independent of rates
This allows us to return to a simpler caching strategy, as all the values
returned by transforms are only resolved during the simulator iterations.
The math done by InflatingInt to compute fractions and products of two
InflatingInts does all its math with the initial values, which introduces
some rounding errors. It may be possible to address these before finalizing
this PR.