[holdings] Pass growth rates to #simulate
This changes the DSL to better differentiate between the simulator
configuration and the running context. I was also able to remove the
@growth_rates instance variable in Simulator, which will prevent
any transforms from accidentally caching it.

If we add additional context to each run besides growth rates, I'd
suggest we refactor growth rates into/up to a Context object.